Python/requirements.txt
protobuf
paho-mqtt
python-dotenv